+++ mobile/src/reducers/auth/auth.reducer.ts
@@ -1,16 +1,14 @@
 import { combineEpics } from 'redux-observable';
-//import { jwtTokenEpic$, listenerAuthStateEpic$, startLoginProcessEpic$, startLogoutProcessEpic$ } from './auth.epics';
-import { startLoginProcessEpic$, shouldRefreshEpic$, jwtTokenEpic$ } from './auth.epics';
+import { shouldRefreshEpic$, jwtTokenEpic$ } from './auth.epics';
 import { AuthActions, AuthActionType } from './auth.action';
 import { JwtToken } from '../../auth/jwt-token-handler';
 import { startLogoutProcessEpic$ } from './logout.epics';
+import { startLoginProcessEpic$ } from './login.epics';
 
 export const authEpics$ = combineEpics(
-    //appStartCheckEpic$,
     startLoginProcessEpic$,
     startLogoutProcessEpic$,
     shouldRefreshEpic$,
-    //listenerAuthStateEpic$ as any,
     jwtTokenEpic$
 );
